Win32 App for shuffling mp3 files on USB drive
This is a simple application used to shuffle mp3 music files on USB
FAT32 partition through reordering the layout of FAT32 directory entry
on USB partition root path.
The purpose is to create a new playlist for some hardware mp3 players
that only have a default in order playback of music files on USB device.